# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telos Wallet API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface UserProfile {
  id: number;
  clerkId: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  role: UserProfileRole;
  twoFactorEnabled: boolean;
  createdAt: string;
}

export interface TotpSetup {
  secret: string;
  otpauthUrl: string;
  qrCodeDataUrl: string;
}

export interface TotpVerifyInput {
  code: string;
}

export interface TotpVerifyResult {
  valid: boolean;
}

export type WalletNetwork = typeof WalletNetwork[keyof typeof WalletNetwork];


export const WalletNetwork = {
  mainnet: 'mainnet',
  testnet: 'testnet',
} as const;

export interface Wallet {
  id: number;
  userId: number;
  label: string;
  zeroAddress: string;
  evmAddress: string;
  network: WalletNetwork;
  createdAt: string;
}

export type WalletInputNetwork = typeof WalletInputNetwork[keyof typeof WalletInputNetwork];


export const WalletInputNetwork = {
  mainnet: 'mainnet',
  testnet: 'testnet',
} as const;

export interface WalletInput {
  label: string;
  network?: WalletInputNetwork;
}

export interface WalletBalance {
  zeroBalance: string;
  evmBalance: string;
  currency: string;
  /** @nullable */
  zeroStaked?: string | null;
  /** @nullable */
  zeroRefunding?: string | null;
}

export interface ReceiveInfo {
  walletId: number;
  zeroAddress: string;
  evmAddress: string;
  zeroQrData?: string;
  evmQrData?: string;
}

export type SendInputNetwork = typeof SendInputNetwork[keyof typeof SendInputNetwork];


export const SendInputNetwork = {
  zero: 'zero',
  evm: 'evm',
} as const;

export interface SendInput {
  fromWalletId: number;
  toAddress: string;
  amount: string;
  /** @nullable */
  memo?: string | null;
  network: SendInputNetwork;
  totpCode: string;
}

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  failed: 'failed',
} as const;

export type TransactionNetwork = typeof TransactionNetwork[keyof typeof TransactionNetwork];


export const TransactionNetwork = {
  zero: 'zero',
  evm: 'evm',
} as const;

export interface Transaction {
  id: number;
  userId: number;
  walletId: number;
  fromAddress: string;
  toAddress: string;
  amount: string;
  currency: string;
  status: TransactionStatus;
  /** @nullable */
  txHash?: string | null;
  /** @nullable */
  memo?: string | null;
  network: TransactionNetwork;
  /** @nullable */
  blockNumber?: number | null;
  createdAt: string;
}

export interface ExplorerTransaction {
  txHash: string;
  fromAddress: string;
  toAddress: string;
  amount: string;
  currency: string;
  blockNumber: number | null;
  timestamp: string;
  status?: string;
  network: string;
}

export interface ExplorerBlock {
  blockNumber: number;
  blockHash: string;
  producer: string;
  txCount: number;
  timestamp: string;
  network: string;
}

export type ExplorerSearchResultType = typeof ExplorerSearchResultType[keyof typeof ExplorerSearchResultType];


export const ExplorerSearchResultType = {
  transaction: 'transaction',
  address: 'address',
  block: 'block',
  notfound: 'notfound',
} as const;

export type ExplorerSearchResultData = { [key: string]: unknown };

export interface ExplorerSearchResult {
  type: ExplorerSearchResultType;
  data: ExplorerSearchResultData;
}

export interface NetworkStats {
  zeroBlockHeight: number;
  evmBlockHeight: number;
  tlosPrice: string;
  totalAccounts: number;
  tps: number;
}

export interface AdminUserList {
  users: UserProfile[];
  total: number;
}

export type AdminUserDetailStatus = typeof AdminUserDetailStatus[keyof typeof AdminUserDetailStatus];


export const AdminUserDetailStatus = {
  active: 'active',
  suspended: 'suspended',
} as const;

export interface AdminUserDetail {
  id: number;
  clerkId: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  role: string;
  twoFactorEnabled: boolean;
  status: AdminUserDetailStatus;
  walletCount: number;
  transactionCount: number;
  createdAt: string;
}

export type AdminUserUpdateRole = typeof AdminUserUpdateRole[keyof typeof AdminUserUpdateRole];


export const AdminUserUpdateRole = {
  user: 'user',
  admin: 'admin',
} as const;

export type AdminUserUpdateStatus = typeof AdminUserUpdateStatus[keyof typeof AdminUserUpdateStatus];


export const AdminUserUpdateStatus = {
  active: 'active',
  suspended: 'suspended',
} as const;

export interface AdminUserUpdate {
  role?: AdminUserUpdateRole;
  status?: AdminUserUpdateStatus;
}

export interface AdminStats {
  totalUsers: number;
  totalWallets: number;
  totalTransactions: number;
  totalVolume: string;
  activeUsersToday: number;
  pendingTransactions: number;
}

export type ListTransactionsParams = {
walletId?: number;
limit?: number;
};

export type GetExplorerTransactionsParams = {
limit?: number;
network?: GetExplorerTransactionsNetwork;
};

export type GetExplorerTransactionsNetwork = typeof GetExplorerTransactionsNetwork[keyof typeof GetExplorerTransactionsNetwork];


export const GetExplorerTransactionsNetwork = {
  zero: 'zero',
  evm: 'evm',
} as const;

export type GetExplorerBlocksParams = {
limit?: number;
network?: GetExplorerBlocksNetwork;
};

export type GetExplorerBlocksNetwork = typeof GetExplorerBlocksNetwork[keyof typeof GetExplorerBlocksNetwork];


export const GetExplorerBlocksNetwork = {
  zero: 'zero',
  evm: 'evm',
} as const;

export type ExplorerSearchParams = {
q: string;
};

export type AdminListUsersParams = {
limit?: number;
offset?: number;
search?: string;
};

export type AdminListTransactionsParams = {
limit?: number;
offset?: number;
};