# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Telos Wallet API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List user transactions
      parameters:
        - name: walletId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
  /transactions/send:
    post:
      operationId: sendTransaction
      tags: [transactions]
      summary: Send TLOS tokens (Zero or EVM)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendInput"
      responses:
        "201":
          description: Transaction submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{txId}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get transaction details
      parameters:
        - name: txId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /explorer/transactions:
    get:
      operationId: getExplorerTransactions
      tags: [explorer]
      summary: Get recent network transactions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: network
          in: query
          schema:
            type: string
            enum: [zero, evm]
      responses:
        "200":
          description: Recent transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExplorerTransaction"
  /explorer/blocks:
    get:
      operationId: getExplorerBlocks
      tags: [explorer]
      summary: Get recent blocks
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: network
          in: query
          schema:
            type: string
            enum: [zero, evm]
      responses:
        "200":
          description: Recent blocks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExplorerBlock"
  /explorer/search:
    get:
      operationId: explorerSearch
      tags: [explorer]
      summary: Search by address or transaction hash
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExplorerSearchResult"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin only)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
  /admin/users/{userId}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Get user details (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update user status (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
  /admin/transactions:
    get:
      operationId: adminListTransactions
      tags: [admin]
      summary: List all transactions (admin only)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: All transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, clerkId, email, role, twoFactorEnabled, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        displayName:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        twoFactorEnabled:
          type: boolean
        createdAt:
          type: string
...
    SendInput:
      type: object
      required: [fromWalletId, toAddress, amount, network, totpCode]
      properties:
        fromWalletId:
          type: integer
        toAddress:
          type: string
        amount:
          type: string
        memo:
          type: ["string", "null"]
        network:
          type: string
          enum: [zero, evm]
        totpCode:
          type: string
    Transaction:
      type: object
      required: [id, userId, walletId, fromAddress, toAddress, amount, currency, status, network, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        walletId:
          type: integer
        fromAddress:
          type: string
        toAddress:
          type: string
        amount:
          type: string
        currency:
          type: string
        status:
          type: string
          enum: [pending, confirmed, failed]
        txHash:
          type: ["string", "null"]
        memo:
          type: ["string", "null"]
        network:
          type: string
          enum: [zero, evm]
        blockNumber:
          type: ["integer", "null"]
        createdAt:
          type: string
    ExplorerTransaction:
      type: object
      required: [txHash, fromAddress, toAddress, amount, currency, blockNumber, timestamp, network]
      properties:
        txHash:
          type: string
        fromAddress:
          type: string
        toAddress:
          type: string
        amount:
          type: string
        currency:
          type: string
        blockNumber:
          type: integer
          nullable: true
        timestamp:
          type: string
        status:
          type: string
        network:
          type: string
    ExplorerBlock:
      type: object
      required: [blockNumber, blockHash, producer, txCount, timestamp, network]
      properties:
        blockNumber:
          type: integer
        blockHash:
          type: string
        producer:
          type: string
        txCount:
          type: integer
        timestamp:
          type: string
        network:
          type: string
    ExplorerSearchResult:
      type: object
      required: [type, data]
      properties:
        type:
          type: string
          enum: [transaction, address, block, notfound]
        data:
          type: object
...
    AdminUserList:
      type: object
      required: [users, total]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total:
          type: integer
    AdminUserDetail:
      type: object
      required: [id, clerkId, email, role, twoFactorEnabled, status, walletCount, transactionCount, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        displayName:
          type: ["string", "null"]
        role:
          type: string
        twoFactorEnabled:
          type: boolean
        status:
          type: string
          enum: [active, suspended]
        walletCount:
          type: integer
        transactionCount:
          type: integer
        createdAt:
          type: string
    AdminUserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, suspended]
...