# lib/api-zod/src/generated/types/explorerSearchResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telos Wallet API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExplorerSearchResultData } from './explorerSearchResultData';
import type { ExplorerSearchResultType } from './explorerSearchResultType';

export interface ExplorerSearchResult {
  type: ExplorerSearchResultType;
  data: ExplorerSearchResultData;
}