# lib/api-zod/src/generated/types/sendInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telos Wallet API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SendInputNetwork } from './sendInputNetwork';

export interface SendInput {
  fromWalletId: number;
  toAddress: string;
  amount: string;
  /** @nullable */
  memo?: string | null;
  network: SendInputNetwork;
  totpCode: string;
}